Module simple_raft::message

Raft message types for sending between nodes.

This module provides data types for messages to be sent between Raft nodes. The top-level message type is RaftMessage. Protobuf-based serialization of all types in this module is provided through the prost crate if the corresponding feature is enabled.

Structs

AppendRequest

A request to append entries to a Raft node’s log.

AppendResponse

The response to an AppendRequest allowing or denying an append to the Raft node’s log.

LogEntry

An entry in a Raft log.

LogIndex

A 1-based index into a Raft log.

RaftMessage

A message sent between Raft nodes.

SendableRaftMessage

A RaftMessage to be sent to a destination.

TermId

The unique, monotonically-increasing ID for a term of Raft group leadership.

VoteRequest

A request to obtain leadership amongst Raft nodes.

VoteResponse

The response to a VoteRequest granting or denying leadership.

Enums

RaftMessageDestination

The destination for a SendableRaftMessage.

Rpc

A Remote Procedure Call message to a Raft node.